Code Refactor.
