+++ network_builder/cli/__init__.py
@@ -1,3 +1,4 @@
 from .cli import CLI
 from . import run
+
 #from .build_transit_segments_parallel import *
